+++ test/spec_lint.rb
@@ -37,6 +37,9 @@ describe Rack::Lint do
     lambda { Rack::Lint.new(valid_app).call({}.freeze) }.must_raise(Rack::Lint::LintError).
       message.must_match(/env should not be frozen, but is/)
 
+    lambda { Rack::Lint.new(valid_app).call({a: 1}) }.must_raise(Rack::Lint::LintError).
+      message.must_equal("env contains non-string keys: [:a]")
+
     lambda {
       e = env
       e.delete("REQUEST_METHOD")